+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/utils/auth-request-adaptors.ts
@@ -152,7 +152,7 @@ const socialProviderMap = (
   const result: SocialProviderResult = {
     authProvidersUserPool,
     ...socialConfigMap,
-    ...userPoolProviders(authProvidersUserPool, { requiredAttributes, ...socialConfigMap }),
+    ...userPoolProviders(authProvidersUserPool, { requiredAttributes, ...socialConfigMap, hostedUI: true }),
   };
   return result;
 };