+++ packages/amplify-cli/src/commands/pull.ts
@@ -1,20 +1,30 @@
 import { pullBackend } from '../pull-backend';
+import { preDeployPullBackend } from '../pre-deployment-pull';
 import { attachBackend } from '../attach-backend';
 import { constructInputParams } from '../amplify-service-helper';
 import { run as envCheckout } from './env/checkout';
 import { stateManager } from 'amplify-cli-core';
-import { get } from 'lodash';
+import _ from 'lodash';
 
 export const run = async context => {
   const inputParams = constructInputParams(context);
   const projectPath = process.cwd();
+
+  if (inputParams.sandboxId) {
+    try {
+      await preDeployPullBackend(context, inputParams.sandboxId);
+    } catch (e) {
+      context.print.error(`Failed to pull sandbox app: ${e.message || 'An unknown error occurred.'}`);
+    }
+    return;
+  }
 
   if (stateManager.currentMetaFileExists(projectPath)) {
     const { appId: inputAppId, envName: inputEnvName } = inputParams.amplify;
     const teamProviderInfo = stateManager.getTeamProviderInfo(projectPath);
     const { envName } = stateManager.getLocalEnvInfo(projectPath);
 
-    const appId = get(teamProviderInfo, [envName, 'awscloudformation', 'AmplifyAppId'], false);
+    const appId = _.get(teamProviderInfo, [envName, 'awscloudformation', 'AmplifyAppId'], false);
 
     const localEnvNames = Object.keys(teamProviderInfo);