+++ packages/amplify-codegen-appsync-model-plugin/src/plugin.ts
@@ -7,6 +7,7 @@ import { AppSyncJSONVisitor } from './visitors/appsync-json-metadata-visitor';
 import { AppSyncModelJavaVisitor } from './visitors/appsync-java-visitor';
 import { AppSyncModelTypeScriptVisitor } from './visitors/appsync-typescript-visitor';
 import { AppSyncModelJavascriptVisitor } from './visitors/appsync-javascript-visitor';
+import { AppSyncModelDartVisitor } from './visitors/appsync-dart-visitor';
 export const plugin: PluginFunction<RawAppSyncModelConfig> = (
   schema: GraphQLSchema,
   rawDocuments: Types.DocumentFile[],
@@ -15,14 +16,12 @@ export const plugin: PluginFunction<RawAppSyncModelConfig> = (
   let visitor;
   switch (config.target) {
     case 'swift':
-    case 'ios':
       visitor = new AppSyncSwiftVisitor(schema, config, {
         selectedType: config.selectedType,
         generate: config.generate,
       });
       break;
     case 'java':
-    case 'android':
       visitor = new AppSyncModelJavaVisitor(schema, config, {
         selectedType: config.selectedType,
         generate: config.generate,
@@ -36,6 +35,12 @@ export const plugin: PluginFunction<RawAppSyncModelConfig> = (
       break;
     case 'javascript':
       visitor = new AppSyncModelJavascriptVisitor(schema, config, {});
+      break;
+    case 'dart':
+      visitor = new AppSyncModelDartVisitor(schema, config, {
+        selectedType: config.selectedType,
+        generate: config.generate,
+      });
       break;
     default:
       return '';