+++ packages/amplify-codegen/src/callbacks/prePushUpdateCallback.js
@@ -3,6 +3,9 @@ const constants = require('../constants');
 const loadConfig = require('../codegen-config');
 const askShouldUpdateCode = require('../walkthrough/questions/updateCode');
 const askShouldUpdateDocs = require('../walkthrough/questions/updateDocs');
+const path = require('path');
+const { isDataStoreEnabled } = require('graphql-transformer-core');
+const { pathManager } = require('amplify-cli-core');
 
 async function prePushUpdateCallback(context, resourceName) {
   const config = loadConfig(context);
@@ -29,6 +32,7 @@ async function prePushUpdateCallback(context, resourceName) {
 
     if (shouldGenerateCode) {
       return {
+        shouldGenerateModels: await isDataStoreEnabled(path.join(pathManager.getBackendDirPath(), 'api', resourceName)),
         gqlConfig: project,
         shouldGenerateDocs,
       };