+++ packages/amplify-frontend-ios/event-handlers/handle-PostInit.js
@@ -1,0 +1,12 @@
+const amplifyApp = require('amplify-app');
+
+async function run(context) {
+  const { frontend } = context.exeInfo.projectConfig;
+  if (frontend === 'ios') {
+    await amplifyApp.run({ skipEnvCheck: true, platform: frontend, skipInit: true, internalOnlyIosCallback: true });
+  }
+}
+
+module.exports = {
+  run,
+};