+++ packages/amplify-provider-awscloudformation/src/push-resources.js
@@ -24,6 +24,7 @@ const { stateManager } = require('amplify-cli-core');
 const constants = require('./constants');
 const { createEnvLevelConstructs } = require('./utils/env-level-constructs');
 const { NETWORK_STACK_LOGICAL_ID } = require('./network/stack');
+const { isAmplifyAdminApp } = require('./utils/admin-helpers');
 
 // keep in sync with ServiceName in amplify-category-function, but probably it will not change
 const FunctionServiceNameLambdaLayer = 'LambdaLayer';
@@ -92,7 +93,7 @@ async function run(context, resourceDefinition) {
 
     // We do not need CloudFormation update if only syncable resources are the changes.
     if (resourcesToBeCreated.length > 0 || resourcesToBeUpdated.length > 0 || resourcesToBeDeleted.length > 0) {
-      await updateCloudFormationNestedStack(context, formNestedStack(context, projectDetails), resourcesToBeCreated, resourcesToBeUpdated);
+      await updateCloudFormationNestedStack(context, await formNestedStack(context, projectDetails), resourcesToBeCreated, resourcesToBeUpdated);
     }
 
     await postPushGraphQLCodegen(context);
@@ -237,7 +238,7 @@ async function updateStackForAPIMigration(context, category, resourceName, optio
       }),
     )
     .then(() => updateS3Templates(context, resources, projectDetails.amplifyMeta))
-    .then(() => {
+    .then(async() => {
       if (!isCLIMigration) {
         spinner.start();
       }
@@ -250,11 +251,11 @@ async function updateStackForAPIMigration(context, category, resourceName, optio
         if (isReverting && isCLIMigration) {
           // When this is a CLI migration and we are rolling back, we do not want to inject
           // an [env] for any templates.
-          nestedStack = formNestedStack(context, projectDetails, category, resourceName, 'AppSync', true);
+          nestedStack = await formNestedStack(context, projectDetails, category, resourceName, 'AppSync', true);
         } else if (isCLIMigration) {
-          nestedStack = formNestedStack(context, projectDetails, category, resourceName, 'AppSync');
+          nestedStack = await formNestedStack(context, projectDetails, category, resourceName, 'AppSync');
         } else {
-          nestedStack = formNestedStack(context, projectDetails, category);
+          nestedStack = await formNestedStack(context, projectDetails, category);
         }
         return updateCloudFormationNestedStack(context, nestedStack, resourcesToBeCreated, resourcesToBeUpdated);
       }
@@ -542,13 +543,19 @@ function uploadTemplateToS3(context, resourceDir, cfnFile, category, resourceNam
 }
 
 /* eslint-disable */
-function formNestedStack(context, projectDetails, categoryName, resourceName, serviceName, skipEnv) {
+async function formNestedStack(context, projectDetails, categoryName, resourceName, serviceName, skipEnv) {
   /* eslint-enable */
   const initTemplateFilePath = path.join(__dirname, '..', 'resources', 'rootStackTemplate.json');
   const nestedStack = context.amplify.readJsonFile(initTemplateFilePath);
   // Track Amplify Console generated stacks
-  if (!!process.env.CLI_DEV_INTERNAL_DISABLE_AMPLIFY_APP_DELETION) {
-    nestedStack.Description = 'Root Stack for AWS Amplify Console';
+  try {
+    const amplifyMeta = stateManager.getMeta();
+    const appId = amplifyMeta.providers[providerName].AmplifyAppId;
+    if(await isAmplifyAdminApp(appId)) {
+      nestedStack.Description = 'Root Stack for AWS Amplify Console';
+    }
+  } catch (err) {
+    console.info('App not deployed yet.');
   }
 
   const { amplifyMeta } = projectDetails;