+++ packages/amplify-e2e-tests/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.31.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@2.30.2...amplify-e2e-tests@2.31.0) (2020-11-30)
+
+
+### Features
+
+* pre-deploy pull, new login mechanism and pkg cli updates ([#5941](https://github.com/aws-amplify/amplify-cli/issues/5941)) ([7274251](https://github.com/aws-amplify/amplify-cli/commit/7274251faadc1035acce5f44699b172e10e2e67d))
+
+
+
+
+
+## [2.30.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@2.30.1...amplify-e2e-tests@2.30.2) (2020-11-28)
+
+
+### Reverts
+
+* "fix: remove app client secret as best practice" ([#5992](https://github.com/aws-amplify/amplify-cli/issues/5992)) ([d7d7fcf](https://github.com/aws-amplify/amplify-cli/commit/d7d7fcf65fb2928f5d97c2ada9fac8ebf3522ee0)), closes [#5731](https://github.com/aws-amplify/amplify-cli/issues/5731) [#5829](https://github.com/aws-amplify/amplify-cli/issues/5829)
+
+
+
+
+
+## [2.30.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@2.30.0...amplify-e2e-tests@2.30.1) (2020-11-27)
+
+**Note:** Version bump only for package amplify-e2e-tests
+
+
+
+
+
+# [2.30.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@2.29.1...amplify-e2e-tests@2.30.0) (2020-11-26)
+
+
+### Features
+
+* **amplify-frontend-flutter:** adds api construction to flutter ([#5955](https://github.com/aws-amplify/amplify-cli/issues/5955)) ([3d69c12](https://github.com/aws-amplify/amplify-cli/commit/3d69c12a26001cc3d9d9be10dee5520797ff7602))
+
+
+
+
+
+## [2.29.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@2.29.0...amplify-e2e-tests@2.29.1) (2020-11-24)
+
+**Note:** Version bump only for package amplify-e2e-tests
+
+
+
+
+
 # [2.29.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@2.27.3...amplify-e2e-tests@2.29.0) (2020-11-22)